Move makeContainerRuntimeWithAgentScheduler out of container-runtime package
