+++ packages/framework/deprecated-agent-scheduler-container-runtime/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@fluidframework/deprecated-aqueduct",
+  "name": "@fluidframework/deprecated-agent-scheduler-container-runtime",
   "version": "0.38.0",
-  "description": "A set of deprecated, backwards-compatible versions of aqueduct classes.",
+  "description": "Deprecated utilities for back-compat with container-runtime with built-in agent-scheduler.",
   "homepage": "https://fluidframework.com",
   "repository": "https://github.com/microsoft/FluidFramework",
   "license": "MIT",
@@ -57,10 +57,12 @@
     "temp-directory": "nyc/.nyc_output"
   },
   "dependencies": {
+    "@fluidframework/agent-scheduler": "^0.38.0",
     "@fluidframework/aqueduct": "^0.38.0",
     "@fluidframework/container-definitions": "^0.38.0",
     "@fluidframework/container-runtime": "^0.38.0",
     "@fluidframework/container-runtime-definitions": "^0.38.0",
+    "@fluidframework/core-interfaces": "^0.38.0",
     "@fluidframework/request-handler": "^0.38.0",
     "@fluidframework/runtime-definitions": "^0.38.0",
     "@fluidframework/synthesize": "^0.38.0",
