--- packages/framework/deprecated-aqueduct/src/agent-scheduler-container-runtime-factories/index.ts
@@ -1,7 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export * from "./agentSchedulerBaseContainerRuntimeFactory";
-export * from "./agentSchedulerContainerRuntimeFactoryWithDefaultDataStore";
