--- packages/framework/deprecated-aqueduct/src/index.ts
@@ -1,13 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-/**
- * The `deprecated-aqueduct` package is a library which contains deprecated, backwards-compatible versions of certain
- * aqueduct classes.  These will not be supported long-term but are meant to ease transitions due to breaking changes.
- *
- * @packageDocumentation
- */
-
-export * from "./agent-scheduler-container-runtime-factories";
