+++ packages/runtime/container-runtime/package.json
@@ -55,7 +55,6 @@
     "temp-directory": "nyc/.nyc_output"
   },
   "dependencies": {
-    "@fluidframework/agent-scheduler": "^0.38.0",
     "@fluidframework/common-definitions": "^0.19.1",
     "@fluidframework/common-utils": "^0.29.0-0",
     "@fluidframework/container-definitions": "^0.38.0",
