+++ packages/test/test-version-utils/package.json
@@ -55,6 +55,7 @@
     "@fluidframework/core-interfaces": "^0.38.0",
     "@fluidframework/counter": "^0.38.0",
     "@fluidframework/datastore-definitions": "^0.38.0",
+    "@fluidframework/deprecated-agent-scheduler-container-runtime": "^0.38.0",
     "@fluidframework/driver-definitions": "^0.38.0",
     "@fluidframework/ink": "^0.38.0",
     "@fluidframework/map": "^0.38.0",
